gh-101 Fix the regex for link replacement in previews
When the Mauro UI will insert a hyperlink to an internal dictionary item, the anchor tag will include a "title" attribute (possibly others). The previous regex would not match on this kind of content.

Adjusted to still find the href (group 1) and the inner text/label (group 2) but also account for additional HTML attributes