gh-114 Fix change papers returning false changes
The issue was because of HTML strings being compared that may have unnecessary whitespace on one side only, but the whitespace is only between tags and not part of the readable content.

Basic string comparison produced false positives, so comparison has to be more intelligent at the cost of speed.